[sil] Move mark_unresolved_move_addr in SILNodes.def so it is not classified as a SingleValueInstruction.
The specific problem here is that this causes MarkUnresolvedMoveAddr to be
within the SINGLE_VALUE_INST_RANGE which defines the classify method used to
determine if casting to a SingleValueInstruction is ok. This is of course not ok
in this case since mark_unresolved_move_addr is actually a NonValueInst.

I noticed this b/c I made the same mistake with a new instruction I am added and
hit a crash in the tests as a result of it being classified as a single value
instruction. After fixing that, I checked for any more instances of this issue
and found that mark_unresolved_move_addr was similarly afflicted.