Flush build step log messages in `build_script_invocation.py` (#77324)
This ensures that log messages are printed as soon as a corresponding build step starts, instead of printing these messages in the end out of order.

For example, before for SwiftPM

```
--- Cleaning swiftpm ---
--- Building swiftpm ---
--- Running tests for swiftpm ---
--- Finished tests for swiftpm ---
--- Installing swiftpm ---
```

was printed at the end of the CI job log with no actual build phase output in between these markers. Now these build phase markers are printed in the correct order and one can infer which log messages were printed for each build phase.