Hook new build into existing Swift project
Setting `SWIFT_ENABLE_NEW_RUNTIME_BUILD` to `ON` will build and run the
new standard library build in addition to using the old build. This is
primarily for basic smoke testing at the moment. Once we're confident,
we can start running tests against the stdlib built with the new
build system.

The mechanism will build the stdlib for each of the same platforms that
the old build system would build for, but emitting a separate build
directory with a separate CMake invocation for each configuration.

This will use the just-built compiler if the compiler was not
cross-compiled. If the compiler were cross-compiled, it will use the
compiler from the toolchain that is known to work on the builder.