Emit a relationship between a typealias and the referenced type.
I've opted to use RelationContains, but I'm not really tied to it. I'm open to any other relation if something else works better.

It is helpful to connect the type(s) referenced in a typealias declaration back to the actual typealias that is using the types. They were already emitted as references but didn't have any connection to how they were being used in the file. This is kind of like how a type conforming to a protocol emits a relationship so that the reference can be connected to the declartion.

This change allows analysis and IDE support to definitively verify that a type reference is part of a typealias.