[maccatalyst] Fix availability_maccatalyst.swift
Trying to fix both availability_maccatalyst and
availability_maccatalyst_app_extension. Success in the former, failure
in the former.

It can be an incorrect recreation of the AvailabilityExtras.h missing
pieces, but the logic and the influences of iOS app extensions in
macCatalyst app extensions either requires a lot more code, or I have
not been able to find the right incantation.