Add Actor.runSynchronously() to synchronously wait for a closure to execute on an actor
The operation is executed at the given priority, and is in the same way as
all other work on the actor. If the calling thread manages to take the
actor lock, it will execute jobs up until this operation is completed,
then return. Otherwise, it will block waiting for a different thread to
execute this operation.

This operation is only available for default actors (ones that do not have
custom executors), and will trap if provided with an actor that has a
custom executor. Note that the blocking nature of this operation makes it
prone to deadlock, so it should be used sparingly.